% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unidades.R
\name{CaF.}
\alias{CaF.}
\title{Grados Celsius a grados Farenheit}
\usage{
CaF.(x)
}
\arguments{
\item{x}{Temperatura en grados Celsius}
}
\value{

}
\description{
Permite convertir de grados Celsius a grados Farenheit
}
\examples{
CaF.(30)
# 86
}
