% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unidades.R
\name{CaK.}
\alias{CaK.}
\title{Convierte de grados Celsius a grados Kelvin o absolutos}
\usage{
CaK.(x)
}
\arguments{
\item{x}{Temperatura en grados Celsius}
}
\value{

}
\description{
Convierte de grados Celsius a grados Kelvin o absolutos
}
\examples{
CaK.(30)
# 303.159

}
