% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unidades.R
\name{FaC.}
\alias{FaC.}
\title{Convierte de grados Farenheit a grados Celsius o centígrados}
\usage{
FaC.(x)
}
\arguments{
\item{x}{Temperatura en grados Farenheit}
}
\value{

}
\description{
Convierte de grados Farenheit a grados Celsius o centígrados
}
\examples{
FaC.(70)
# 21.11111

}
